refactor(frontend): move ui and lib into shared layer
