+++ frontend/packages/frontend/MIGRATION.md
@@ -1,0 +1,9 @@
+# Migration Guide
+
+## Shared library and UI paths
+
+- `@/lib/*` -> `@/shared/lib/*`
+- `@/components/ui/*` -> `@/shared/ui/*`
+
+Update imports accordingly.
+
