+++ frontend/packages/frontend/src/components/PhotoFlags.tsx
@@ -1,4 +1,4 @@
-import { Badge } from '@/components/ui/badge';
+import { Badge } from '@/shared/ui/badge';
 
 export type PhotoFlagsProps = {
   isBW?: boolean;
