+++ frontend/packages/frontend/src/components/ScoreBar.tsx
@@ -1,4 +1,4 @@
-import { Label } from "@/components/ui/label";
+import { Label } from "@/shared/ui/label";
 
 interface ScoreBarProps {
     label: string;
