+++ frontend/packages/frontend/src/entities/README.md
@@ -1,0 +1,4 @@
+# Entities
+
+Domain models and related logic.
+
