+++ frontend/packages/frontend/src/pages/admin/UsersPage.tsx
@@ -4,7 +4,7 @@ import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { manageUsersTitle, saveUserButtonText, phoneNumberLabel, telegramLabel } from '@photobank/shared/constants';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/shared/ui/button';
 import {
   Form,
   FormControl,
@@ -12,9 +12,9 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
-import { Textarea } from '@/components/ui/textarea';
+} from '@/shared/ui/form';
+import { Input } from '@/shared/ui/input';
+import { Textarea } from '@/shared/ui/textarea';
 import {
   useGetAdminUsersQuery,
   useUpdateAdminUserMutation,
@@ -45,7 +45,7 @@ function UserEditor({ user, onSave }: UserEditorProps) {
     <div className="border p-4 rounded space-y-2">
       <h2 className="font-semibold">{user.email}</h2>
       <Form {...form}>
-        {/* eslint-disable-next-line @typescript-eslint/no-misused-promises */}
+        { }
         <form onSubmit={form.handleSubmit((d) => onSave(user.id!, d))} className="space-y-4">
           <FormField
             control={form.control}
