+++ frontend/packages/frontend/src/shared/ui/README.md
@@ -1,0 +1,4 @@
+# Shared UI
+
+Reusable UI components based on shadcn/ui.
+
