+++ frontend/packages/frontend/src/shared/ui/aspect-ratio.tsx
