+++ frontend/packages/frontend/src/shared/ui/form.tsx
@@ -11,8 +11,8 @@ import {
   type FieldValues,
 } from "react-hook-form"
 
-import { cn } from "@/lib/utils"
-import { Label } from "@/components/ui/label"
+import { cn } from "@/shared/lib/utils"
+import { Label } from "@/shared/ui/label"
 
 const Form = FormProvider
