+++ frontend/packages/frontend/src/shared/ui/table.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shared/lib/utils"
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
