+++ frontend/packages/frontend/src/widgets/README.md
@@ -1,0 +1,4 @@
+# Widgets
+
+Composite UI blocks built from features and entities.
+
