+++ frontend/packages/frontend/test/utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { cn } from '../src/lib/utils';
+import { cn } from '../src/shared/lib/utils';
 
 describe('cn', () => {
   it('joins class names', () => {
